Clean up integration tests, test for active section
