+++ integration-tests/integration-tests.js
@@ -4,9 +4,8 @@
  * And a11y tests last as they take longer to run
  */
 require('./test-suites/server.test');
-require('./test-suites/locale.test');
+require('./test-suites/core.test');
 require('./test-suites/legacy.test');
-require('./test-suites/sections.test');
 require('./test-suites/survey.test');
 require('./test-suites/materials.test');
 require('./test-suites/user.test');
