+++ integration-tests/test-suites/server.test.js
@@ -25,10 +25,6 @@ describe('Basic server tests', () => {
             .get('/')
             .then(res => {
                 expect(res).to.have.status(200);
-                const metaTitle = `Home | Big Lottery Fund`;
-                expect(res.text).to.include(`<title>${metaTitle}</title>`);
-                expect(res.text).to.include(`<meta name="title" content="${metaTitle}">`);
-                expect(res.text).to.include(`<meta property="og:title" content="${metaTitle}">`);
             });
     });
 
@@ -72,23 +68,6 @@ describe('Basic server tests', () => {
             .then(res => {
                 expect(res).to.have.status(200);
                 expect(res).to.have.header('content-type', /^image\/png/);
-            });
-    });
-
-    it('should allow contrast preferences to be set', () => {
-        let redirectUrl = 'http://www.google.com';
-        return chai
-            .request(server)
-            .get('/contrast/high')
-            .query({
-                url: redirectUrl
-            })
-            .redirects(0)
-            .catch(err => err.response)
-            .then(res => {
-                expect(res).to.have.cookie(config.get('cookies.contrast'));
-                expect(res).to.redirectTo(redirectUrl);
-                expect(res).to.have.status(302);
             });
     });
