+++ views/components/nav.njk
@@ -3,7 +3,7 @@
 {% macro navlinks(itemClass) %}
     {% for id, section in routes %}
         {% set currentSectionId = getCurrentSection(sectionId, pageId)  %}
-        <li class="{% if currentSectionId === id %}is-selected{% endif %}{% if itemClass %} {{ itemClass }}{% if loop.last %} {{ itemClass }}--last{% endif %}{% endif %}">
+        <li class="{% if currentSectionId === id %}is-selected{% endif %}{% if itemClass %} {{ itemClass }}{% if loop.last %} {{ itemClass }}--last{% endif %}{% endif %} qa-nav-link qa-nav-link--{{ id }}">
             <a href="{{ buildUrl(id) }}">{{ __(section.langTitlePath) }}</a>
         </li>
     {% endfor %}
