+++ views/includes/header.njk
@@ -23,7 +23,7 @@
             <nav role="navigation"
                 class="global-header__nav u-desktop-only u-dont-print">
                 <ul class="global-nav">
-                    {{ navlinks('qa-nav-link global-nav__link') }}
+                    {{ navlinks('global-nav__link') }}
                     <li class="global-nav__lang">
                         {{ navLangLink() }}
                     </li>
